0613成果

+++ source/src/main/java/dto/user.java
@@ -7,7 +7,7 @@ public class user implements Serializable {
 	private String user_name;
 	private String family_id;
 	private String password;
-	private double share_goal;
+	private float share_goal;
 
 	public String getUser_id() {
 		return user_id;
@@ -41,15 +41,15 @@ public class user implements Serializable {
 		this.password = password;
 	}
 
-	public double getShare_goal() {
+	public float getShare_goal() {
 		return share_goal;
 	}
 
-	public void setShare_goal(double share_goal) {
+	public void setShare_goal(float share_goal) {
 		this.share_goal = share_goal;
 	}
 
-	public user(String user_id, String user_name, String family_id, String password, double share_goal) {
+	public user(String user_id, String user_name, String family_id, String password, float share_goal) {
 		super();
 		this.user_id = user_id;
 		this.user_name = user_name;
